requirements.txt
django==3.1.6
django-crispy-forms==1.11.0
Pillow==8.1.0
pytest-django==4.1.0
pytest-cov==2.11.1
djangorestframework==3.12.2
